split file into multiple files
