+++ pyspline/__init__.py
@@ -1,6 +1,6 @@
 __version__ = "1.2.1"
 
 from . import pySpline
-from .pySpline import Curve
-from .pySpline import Surface
-from .pySpline import Volume
+from .pyCurve import Curve
+from .pySurface import Surface
+from .pyVolume import Volume
